+++ src/com/android/bluetooth/opp/BluetoothOppReceiveFileInfo.java
@@ -104,7 +104,7 @@ public class BluetoothOppReceiveFileInfo {
             try {
                 if (metadataCursor.moveToFirst()) {
                     hint = metadataCursor.getString(0);
-                    length = metadataCursor.getInt(1);
+                    length = metadataCursor.getLong(1);
                 }
             } finally {
                 metadataCursor.close();
